+++ src/process/processors/mod.rs
@@ -1,5 +1,7 @@
 //! A collection of utility processors that can be used when creating rules.
 
 mod find_identifier;
+mod find_usage;
 
 pub use find_identifier::*;
+pub(crate) use find_usage::*;