+++ tests/rule_tests/mod.rs
@@ -309,5 +309,6 @@ mod remove_method_definition;
 mod remove_nil_declaration;
 mod remove_types;
 mod remove_unused_if_branch;
+mod remove_unused_variable;
 mod remove_unused_while;
 mod rename_variables;